Created new sub-modules `scripts` and added basic `paraproj` command
